image_compare/requirements.txt
requests
pillow
torch
git+https://github.com/openai/CLIP.git